app/requirements.txt
anyio==4.9.0
certifi==2025.4.26
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
google-auth==2.40.3
google-auth-oauthlib==1.2.2
gspread==6.2.1
oauth2client==4.1.3
pandas==2.3.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
python-telegram-bot[job-queue]==22.0
sniffio==1.3.1